pyproject.toml
[project]
name = "tds-ta"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "requests",
    "openai>=1.0.0",
    "html2text",
    "beautifulsoup4",
    "google-genai",
    "httpx",
    "pillow",
    "python-dotenv",
    "tiktoken",
    "numpy",
    "tqdm",
    "google-generativeai",
    "openai",
    "flask>=3.1.1",
    "python-multipart>=0.0.20",
]
